# apps/server/openapi.json
{
  "components": {
    "schemas": {
...
      "AnnounceRequest": {
        "description": "Post an admin announcement into a chat as a centered system_card (官方公告).",
        "properties": {
          "content": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "AnnounceRequest",
        "type": "object"
      },
...
      "BlockListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/BlockedUser"
            },
            "title": "Data",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "data",
          "total"
        ],
        "title": "BlockListResponse",
        "type": "object"
      },
      "BlockUserRequest": {
        "properties": {
          "user_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "BlockUserRequest",
        "type": "object"
      },
...
      "ChatListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ChatSummary"
            },
            "title": "Data",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "data",
          "total"
        ],
        "title": "ChatListResponse",
        "type": "object"
      },
      "ChatMembersResponse": {
        "description": "A chat's members (group roster: resolves sender names + the member panel).",
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ChatParticipant"
            },
            "title": "Data",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "data",
          "total"
        ],
        "title": "ChatMembersResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StartDmRequest": {
        "description": "Open (or reuse) a 1:1 chat with another user (by their user id).",
        "properties": {
          "user_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "StartDmRequest",
        "type": "object"
      },
      "StatusResponse": {
        "properties": {
          "status": {
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "StatusResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Multi-Agent AI Workspace API",
    "title": "AgentCore",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/messages/blocks": {
      "get": {
        "description": "The users this user has blocked.",
        "operationId": "list_blocks_v1_messages_blocks_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Blocks",
        "tags": [
          "messages"
        ]
      },
      "post": {
        "description": "Block a user (symmetric: severs DMs and hides each from the other's search).\n\n422 self-block; 404 unknown target.",
        "operationId": "block_user_v1_messages_blocks_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Block User",
        "tags": [
          "messages"
        ]
      }
    },
    "/v1/messages/blocks/{target_id}": {
      "delete": {
        "description": "Remove a user from the block list (idempotent).",
        "operationId": "unblock_user_v1_messages_blocks__target_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "target_id",
            "required": true,
            "schema": {
              "title": "Target Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unblock User",
        "tags": [
          "messages"
        ]
      }
    },
    "/v1/messages/chats": {
      "get": {
        "description": "This user's chat list (recent first), with unread counts and dm peers.",
        "operationId": "list_chats_v1_messages_chats_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Chats",
        "tags": [
          "messages"
        ]
      }
    },
    "/v1/messages/chats/dm": {
      "post": {
        "description": "Open (or reuse) a 1:1 chat with another user (by their user id).\n\n422 self-dm; 404 unknown/disabled peer; 403 when blocked or the peer only\naccepts contacts. The peer joins as a pending message request until they reply.",
        "operationId": "start_dm_v1_messages_chats_dm_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartDmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Dm",
        "tags": [
          "messages"
        ]
      }
    },
    "/v1/messages/chats/{chat_id}/announce": {
      "post": {
        "description": "Post an admin announcement as a centered system_card, fanned out to members\n(platform-admin only). 404 unknown chat; 422 for a dm.",
        "operationId": "announce_v1_messages_chats__chat_id__announce_post",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "title": "Chat Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnounceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Announce",
        "tags": [
          "messages"
        ]
      }
    },
    "/v1/messages/chats/{chat_id}/leave": {
      "post": {
        "description": "Leave a group/official chat. 404 if not a member; 422 for a dm (can't leave).",
        "operationId": "leave_chat_v1_messages_chats__chat_id__leave_post",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "title": "Chat Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Leave Chat",
        "tags": [
          "messages"
        ]
      }
    },
    "/v1/messages/chats/{chat_id}/members": {
      "get": {
        "description": "A chat's members — the group roster (resolves sender names + member panel).\n\n404 if the requester is not a member (IDOR-safe, no existence leak). Each row\ncarries the member's platform-admin and admin-mute flags for the panel.",
        "operationId": "list_chat_members_v1_messages_chats__chat_id__members_get",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "title": "Chat Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMembersResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Chat Members",
        "tags": [
          "messages"
        ]
      }
    },
    "/v1/messages/chats/{chat_id}/members/{target_id}": {
      "delete": {
        "description": "Remove a member from a group (platform-admin only); posts a system notice.\n\n403 non-admin (AdminUser gate); 404 unknown chat or non-member target; 422 for\na dm; 403 when the target is an admin (admins can't be moderated).",
        "operationId": "kick_member_v1_messages_chats__chat_id__members__target_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "title": "Chat Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "target_id",
            "required": true,
            "schema": {
              "title": "Target Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Kick Member",
        "tags": [
          "messages"
        ]
      }
    },
    "/v1/messages/chats/{chat_id}/members/{target_id}/mute": {
      "post": {
        "description": "Mute / unmute a member (platform-admin only): a muted member can read but\nnot send (403 on send). Same gates as kick.",
        "operationId": "mute_member_v1_messages_chats__chat_id__members__target_id__mute_post",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "title": "Chat Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "target_id",
            "required": true,
            "schema": {
              "title": "Target Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminMuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mute Member",
        "tags": [
          "messages"
        ]
      }
    },
    "/v1/messages/chats/{chat_id}/membership": {
      "patch": {
        "description": "Patch this user's per-chat flags (mute / pin). 404 if not a member.",
        "operationId": "update_membership_v1_messages_chats__chat_id__membership_patch",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "title": "Chat Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMembershipRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Membership",
        "tags": [
          "messages"
        ]
      }
    },
...
    "/v1/messages/chats/{chat_id}/read": {
      "post": {
        "description": "Advance this user's read cursor (drives unread counts). 404 if not a member.",
        "operationId": "mark_chat_read_v1_messages_chats__chat_id__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "title": "Chat Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkReadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark Chat Read",
        "tags": [
          "messages"
        ]
      }
    },
...
  }
}